AutoLogin

+++ Frontend/lib/model/addUser.dart
@@ -7,7 +7,7 @@ class User {
   String email;
   String gender;
   bool driverMode;
-  double balance;
+  int balance;
   String apiKey;
   User(
       {this.id,
@@ -29,7 +29,7 @@ class User {
       gender: user['gender'],
       driverMode: user['driver_mode'],
       apiKey: user['api_key']);
-  void setBalnce(double _balance) {
+  void setBalnce(int _balance) {
     balance = _balance;
   }
 }
